+++ CHANGES.md
@@ -7,6 +7,9 @@
 + build: update all dependencies to fix check-docs ci failure
 + core: Increase MSRV to 1.91.1 to incorporate latest Rust features
 + build: Use `workspace.dependencies` to unify dependencies version management across all crates in project's workspace.
++ all: Switch from `once_cell`'s `Lazy`, `sync::OnceCell`, `unsync::OnceCell`
+  to `std` types `LazyLock`, `sync::OnceLock` and `cell::OnceCell`.
+  Remove `once_cell` dependency from `Cargo.toml` files.
 
 ## 0.10.0 - 2025-09-01